requirements.txt
flask== 1.1.1
sqlite3==2.6.0
pickle==4.0
sklearn==0.21.3
tensorflow==2.0

